requirements.txt
# Charm Scaffolding Automation System Dependencies
# Install with: pip install -r requirements.txt

# OpenAI API client
openai>=1.0.0

# YAML parsing
pyyaml>=6.0

# CLI framework
click>=8.0

# Data validation
pydantic>=2.0

# Retry logic for API calls
tenacity>=8.0

# Environment variable management
python-dotenv>=1.0

# Optional: For development and testing
# pytest>=7.0
# pytest-cov>=4.0
# black>=23.0
# ruff>=0.1.0

